fixed circular dependencies
