+++ packages/core/src/index.ts
@@ -1,13 +1,14 @@
 import { Agile } from './agile';
 
 export * from './agile';
-export * from './shared';
 export * from './utils';
+export * from './logCodeManager';
 export * from './integrations';
 export * from './storages';
 export * from './runtime';
 export * from './state';
 export * from './collection';
 export * from './computed';
+export * from './shared';
 
 export default Agile;
